third_party/blink/web_tests/svg/css/stroke-opacity-change-no-layout-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS internals.needsLayoutCount() is 0
Change of stroke-opacity should cause style recalc only, no layout. 
